% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TIMBR_source.R, R/TIMBR_source_bk.R,
%   R/TIMBR_source_bk2.R, R/TIMBR_source_bk3.R
\name{additive.design}
\alias{additive.design}
\title{Create an additive genetic design matrix}
\usage{
additive.design(J, type)

additive.design(J, type)

additive.design(J, type)

additive.design(J, type)
}
\arguments{
\item{J}{number of haplotypes}

\item{type}{format of the diplotype matrix, currently only supports "happy"}

\item{J}{number of haplotypes}

\item{type}{format of the diplotype matrix, currently only supports "happy"}

\item{J}{number of haplotypes}

\item{type}{format of the diplotype matrix, currently only supports "happy"}

\item{J}{number of haplotypes}

\item{type}{format of the diplotype matrix, currently only supports "happy"}
}
\value{
matrix of (half) haplotype dosages corresponding to each diplotype state

matrix of (half) haplotype dosages corresponding to each diplotype state

matrix of (half) haplotype dosages corresponding to each diplotype state

matrix of (half) haplotype dosages corresponding to each diplotype state
}
\description{
Returns the additive design matrix for the number of specified haplotypes, in the specified format.

Returns the additive design matrix for the number of specified haplotypes, in the specified format.

Returns the additive design matrix for the number of specified haplotypes, in the specified format.

Returns the additive design matrix for the number of specified haplotypes, in the specified format.
}
\examples{
additive.design(8, "happy")

additive.design(8, "happy")

additive.design(8, "happy")

additive.design(8, "happy")

}
