% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TIMBR_source.R
\name{approx.alpha.mle}
\alias{approx.alpha.mle}
\title{#' @keywords internal
TIMBR.scan.plot <- function(TIMBR.scan.output, col="black", add=F, ylim=NULL){
  plot.data <- c(TIMBR.scan.output$intervals[,1], TIMBR.scan.output$intervals[,2]-1)
  plot.data <- cbind(plot.data, c(TIMBR.scan.output$ln.BFs, TIMBR.scan.output$ln.BFs))
  plot.data <- plot.data[order(plot.data[,1]),]
  
  if (add){
    lines(plot.data[,1], plot.data[,2], col=col)
  } else {
    if (is.null(ylim)){
      plot(plot.data[,1], plot.data[,2], type="l", col=col, ylab="lnBF", xlab="Location")
    } else {
      plot(plot.data[,1], plot.data[,2], type="l", col=col, ylab="lnBF", xlab="Location", ylim=ylim)
    }
  }
}}
\usage{
approx.alpha.mle(TIMBR.output, ln.BF.full = NULL)
}
\description{
#' @keywords internal
TIMBR.scan.plot <- function(TIMBR.scan.output, col="black", add=F, ylim=NULL){
  plot.data <- c(TIMBR.scan.output$intervals[,1], TIMBR.scan.output$intervals[,2]-1)
  plot.data <- cbind(plot.data, c(TIMBR.scan.output$ln.BFs, TIMBR.scan.output$ln.BFs))
  plot.data <- plot.data[order(plot.data[,1]),]
  
  if (add){
    lines(plot.data[,1], plot.data[,2], col=col)
  } else {
    if (is.null(ylim)){
      plot(plot.data[,1], plot.data[,2], type="l", col=col, ylab="lnBF", xlab="Location")
    } else {
      plot(plot.data[,1], plot.data[,2], type="l", col=col, ylab="lnBF", xlab="Location", ylim=ylim)
    }
  }
}
}
\keyword{internal}
