% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/climb.R
\name{climb}
\alias{climb}
\title{Climb a forest from parents to children}
\usage{
climb(.data, ..., .deep = TRUE)
}
\arguments{
\item{.data}{A forest.}

\item{...}{A list of node names to climb the forest.}

\item{.deep}{Whether to search deeply for node names or not?}
}
\value{
A forest.
}
\description{
Climb a forest from parents to children with one or more node names.
}
