% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cluster_trees.R
\name{cluster_trees}
\alias{cluster_trees}
\title{Cluster trees based on their similarityof a random forest}
\usage{
cluster_trees(rf, num.clusters = NULL, distance.matrix = NULL)
}
\arguments{
\item{rf}{Object of class \code{ranger} used with \code{write.forest = TRUE} to select trees from.}

\item{num.clusters}{Number of clusters to seperate the trees in. Can be an integer smaller than the number of trees or "AUTO" to auto detect the best number of clusters}

\item{distance.matrix}{Add matrix of precalculated distances.}
}
\value{
\item{\code{cluster_ids}}{List of numerical vectors where each entry represents the tree ids belonging to one cluster}
}
\description{
\code{cluster_trees} uses pair-wise dissimilarity of trees in a random
forest trained with \code{ranger} to idetify clusters of similar trees.
}
\examples{
require(ranger)
require(timbR)

## Train random forest with ranger
rg.iris <- ranger(Species ~ ., data = iris, write.forest=TRUE, num.trees = 500)

## Calculate pair-wise distances for all trees
distances <- measure_distances(rf = rg.iris, metric = "weighted splitting variables")

## Cluster trees in two clusters
clustered_trees <- cluster_trees(rf = rg.iris, num.clusters = 2, distance.matrix = distances)

## Plot clustered trees colored by cluster in principal component space
plot(clustered_trees$PC1, clustered_trees$PC2, col = clustered_trees$cluster)


}
\author{
Bjoern-Hergen Laabs, M.Sc.
}
