% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dplyr.R
\name{dplyr}
\alias{dplyr}
\alias{mutate.timbr_forest}
\alias{summarise.timbr_forest}
\alias{select.timbr_forest}
\alias{relocate.timbr_forest}
\alias{rows_update.timbr_forest}
\alias{rows_patch.timbr_forest}
\alias{rowwise.timbr_forest}
\alias{ungroup.timbr_forest}
\title{dplyr methods for forest objects}
\usage{
\method{mutate}{timbr_forest}(.data, ...)

\method{summarise}{timbr_forest}(.data, ..., .node = NULL)

\method{select}{timbr_forest}(.data, ...)

\method{relocate}{timbr_forest}(.data, ...)

\method{rows_update}{timbr_forest}(x, y, by = NULL, ...)

\method{rows_patch}{timbr_forest}(x, y, by = NULL, ...)

\method{rowwise}{timbr_forest}(data, ...)

\method{ungroup}{timbr_forest}(x, ...)
}
\arguments{
\item{.data}{A forest.}

\item{...}{Other arguments.}

\item{.node}{`NULL` (default) or a vector to create new nodes.}

\item{x}{A forest.}

\item{y}{A data frame.}

\item{by}{An unnamed character vector giving the key columns.}

\item{data}{A forest.}
}
\value{
A forest.
}
\description{
dplyr methods for forest objects.
}
