% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/timbr-utility.R
\name{duplicateNodes}
\alias{duplicateNodes}
\title{Get duplicate nodes from LumberYard object}
\usage{
duplicateNodes(ly)
}
\arguments{
\item{x}{a \code{\link{lumberYard}}}
}
\value{
a vector of node indices that are duplicates. The first duplicate is
preserved and not included in the index vector.
}
\description{
Returns the index positions of nodes with identical node-text. Much faster
and memory efficient than running \code{which(duplicated(t(nodes))}.
}

