% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/forest.R
\name{forest_by}
\alias{forest_by}
\title{Constructs a forest by one or more variables}
\usage{
forest_by(.data, ...)
}
\arguments{
\item{.data}{A data frame.}

\item{...}{Variables.}
}
\value{
A forest.
}
\description{
\code{forest_by()} constructs a forest by one or more variables.
}
