% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_tree.R
\name{generate_tree}
\alias{generate_tree}
\title{Generate most representative tree for a random forest}
\usage{
generate_tree(rf, metric = "splitting variables", train_data, test_data = NULL)
}
\arguments{
\item{rf}{Object of class \code{ranger} used with \code{write.forest = TRUE} to generate tree for.}

\item{metric}{Specification of the tree metric. Available are "splitting variables",
"weighted splitting variables", "terminal nodes" and "prediction".}

\item{train_data}{Data set for training of artificial representative tree}

\item{...}{Further paramters passed on to measure_distances (e.g. test_data)}
}
\value{
\item{\code{rep.trees}}{\code{ranger} object containing the artificial most representative tree}
}
\description{
\code{generate_tree} uses pair-wise dissimilarity of trees in a random
forest trained with \code{ranger} to generate an artificail most representative tree
which is not part aof the original ensemble
}
\examples{
require(ranger)
require(timbR)

## Train random forest with ranger
rg.iris <- ranger(Species ~ ., data = iris, write.forest=TRUE, num.trees = 10)

## Calculate pair-wise distances for all trees
rep_tree <- generate_tree(rf = rg.iris, metric = "splitting variables", train_data = iris)

}
\author{
Bjoern-Hergen Laabs, M.Sc.
}
