% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_art_prob.R
\name{get_art_prob}
\alias{get_art_prob}
\title{\code{get_art_prob} transforms a classification tree to a probability predictor (scoring classifier), e.g. an artificial representative tree (art)}
\usage{
get_art_prob(art, train_data)
}
\arguments{
\item{art}{Tree object of class \code{ranger}, e.g. an artificial representative tree (art) build with generate_tree()}

\item{train_data}{Additional data set comparable to the data set \code{rf} was build on. Only needed for metric "prediction".}
}
\value{
ranger object predicting probabilities instead of classes
}
\description{
\code{get_art_prob} transforms a classification tree to a probability predictor (scoring classifier), e.g. an artificial representative tree (art)
}
\author{
Lea Louisa Kronziel, M.sc.
}
