% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_brier_score.R
\name{get_brier_score}
\alias{get_brier_score}
\title{Get brier score for predicted probabilities}
\usage{
get_brier_score(predictions, true_class_labels)
}
\arguments{
\item{predictions}{2D object with predictions (data.frame or matrix), observations as rows, classes as columns}

\item{true_class_labels}{Vector with true class labels}
}
\value{
Brier score as numerical value
}
\description{
Get brier score for predicted probabilities
}
\author{
Lea Louisa Kronziel, M.Sc.
}
