% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_calibrated_nodes.R
\name{get_calibrated_nodes}
\alias{get_calibrated_nodes}
\title{Node-wise calibration of predictions of decision tree using mondrian conformal prediction}
\usage{
get_calibrated_nodes(
  y_cal_pred,
  y_cal,
  y_test_pred,
  significance_level,
  tree,
  cal_data,
  test_data,
  dependent_varname,
  calibrate_all_nodes = FALSE
)
}
\arguments{
\item{y_cal_pred}{Predictions of decision tree for calibration set.}

\item{y_cal}{True label of outcome variable for calibration set. Has to be in the same order as y_cal_pred.}

\item{y_test_pred}{Predictions of test data set that should be calibrated.}

\item{significance_level}{Level of uncertainty that should be reached by calibration, should be between 0 and 1.}

\item{tree}{Tree whose predictions are to be calibrated, tree should be an object of class \code{ranger}.}

\item{cal_data}{Data frame with calibration data (should contain same variables as train data).}

\item{test_data}{Data frame with test data (should contain same variables as train data).}

\item{dependent_varname}{Name of the dependent variable used to create the tree.}

\item{calibrate_all_nodes}{If TRUE prediction interval is calculated for all nodes, if FALSE only for termial nodes.}
}
\value{
data frame with point prediction and lower and upper bound of prediction interval for nodes.
}
\description{
\code{get_calibrated_nodes} calculates lower and upper limit for prediction interval for each observation of the test data set
using mondrian conformal prediction
}
\author{
Lea Louisa Kronziel, M.Sc.
}
