% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_calibrated_prediction_regression.R
\name{get_calibrated_prediction_regression}
\alias{get_calibrated_prediction_regression}
\title{Calibration of predictions of regression model using conformal prediction}
\usage{
get_calibrated_prediction_regression(
  y_cal_pred,
  y_cal,
  y_test_pred,
  significance_level
)
}
\arguments{
\item{y_cal_pred}{Predictions of regression model for calibration set.}

\item{y_cal}{True label of outcome variable for calibration set. Has to be in the same order as y_cal_pred.}

\item{y_test_pred}{Predictions of test data set that should be calibrated.}

\item{significance_level}{Level of uncertainty that should be reached by calibration, should be between 0 and 1.}
}
\value{
\item{data frame with point prediction and lower and upper bound of prediction interval.}
}
\description{
\code{get_calibrated_prediction_regression} calibrates regression predictions using conformal prediction
to a point prediction with prediction interval for each observation of the test data set.
}
\examples{
require(ranger)
require(timbR)

regr_data <- longley \%>\% data.frame()
# Train random forest with ranger
rf <- ranger(Employed ~ ., data = regr_data, num.trees = 10, importance = "permutation")

# Calculate pair-wise distances for all trees
rep_tree <- generate_tree_reimplementation(rf = rf, metric = "splitting variables", train_data = regr_data, dependent_varname = "Employed", importance.mode = TRUE, imp.num.var = 2)

# Get predictions
rep_tree_predictions <- predict(rep_tree, regr_data)$predictions

# Calibrated predictions
get_calibrated_prediction_regression(y_cal_pred = rep_tree_predictions, y_cal = regr_data$Employed, y_test_pred = rep_tree_predictions, significance_level = 0.05)

}
\author{
Lea Louisa Kronziel, M.Sc.
}
