% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_cpd_plot_node.R
\name{get_cpd_plot_node}
\alias{get_cpd_plot_node}
\title{Plot conformal predictive systems of a terminal node}
\usage{
get_cpd_plot_node(
  nodeID,
  cal_data_node,
  dependent_varname,
  tree,
  threshold,
  significance_level = 0.05,
  interval_type = "two-tailed",
  direction = NULL,
  show_point_prediction = TRUE,
  show_prediction_interval = TRUE
)
}
\arguments{
\item{nodeID}{ID of terminal node, e.g select one nodeID of a terminal node from ranger::treeInfo()}

\item{cal_data_node}{Calibration data (data frame that includes outcome variable) that reaches in that terminal node}

\item{dependent_varname}{Name of the dependent variable used to create the tree.}

\item{tree}{Tree that should be used to build the CPS plot, tree should be an object of class \code{ranger}.}

\item{threshold}{Vector of numerical threshhold(s) for which the probability of reaching this threshold is to be displayed in the plot; in addition, the probability of reaching at least the threshold is returned. Set to NULL if not used.}

\item{significance_level}{Level of uncertainty that should be reached by calibration, should be between 0 and 1.}

\item{interval_type}{Type of interval, choose either two-tailed or one-tailed}

\item{direction}{Direction of one-tailed interval, choose either left-tailed or right-tailed. Set to NULL if not used.}

\item{show_point_prediction}{Set to TRUE, if you want to display dashed line with point prediction for this terminal node in plot}

\item{show_prediction_interval}{Set to TRUE, if you want to display prediction interval for this terminal node in plot}
}
\value{
\item{list, first element is ggplot object, second a data.frame with probabilities beeing above inserted treshold}
}
\description{
\code{get_cpd_plot_node} builds ggplot object for CPS of a terminal node of a tree
}
\author{
Lea Louisa Kronziel, M.Sc.
}
