% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_factor_split_levels.R
\name{get_factor_split_levels}
\alias{get_factor_split_levels}
\title{Get split criterion (levels) for factor or character variables
Character variables are automatically turned into factors in ranger() from the package "ranger"
Considers the three possibilities to treat and split factors (Parameter "respect.unordered.factors" in ranger()): 
- Treat the factor like an ordinal variable ("none", "NULL" or "FALSE" in ranger())
- Treat the factor as nominal and check all 2-partitions for split ("partition")
- Treat factor as ordinal and determine optimal order of levels for the random forest ("order", "TRUE" in ranger())}
\usage{
get_factor_split_levels(
  node_id,
  split_variable,
  split_variable_type,
  split_value,
  train_data_df,
  rf_list
)
}
\arguments{
\item{node_id}{Node ID of the node whose parent ID is to be determined}

\item{split_variable}{Column name of the split variable in training data}

\item{split_variable_type}{Type of split variable: "factor" for factors, "character" for character variable}

\item{split_value}{Number(s) used for the split, for example the entry "splival" from the function treeInfo()}

\item{train_data_df}{Data frame of the training data with which the random forest was trained}

\item{rf_list}{Random forest, which is built like the one you get from ranger()}
}
\value{
Character with levels passed from the parent node to the considered node
}
\description{
Get split criterion (levels) for factor or character variables
Character variables are automatically turned into factors in ranger() from the package "ranger"
Considers the three possibilities to treat and split factors (Parameter "respect.unordered.factors" in ranger()): 
- Treat the factor like an ordinal variable ("none", "NULL" or "FALSE" in ranger())
- Treat the factor as nominal and check all 2-partitions for split ("partition")
- Treat factor as ordinal and determine optimal order of levels for the random forest ("order", "TRUE" in ranger())
}
