% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_inbag_data.R
\name{get_inbag_data}
\alias{get_inbag_data}
\title{Get inbag data of decision tree}
\usage{
get_inbag_data(rf_list, train_data_df, tree_number)
}
\arguments{
\item{rf_list}{Random forest, which is built like the one you get from ranger()}

\item{train_data_df}{Data frame of the training data with which the random forest was trained}

\item{tree_number}{Number of the decision tree of the rf_list to be displayed}
}
\value{
Data frame with inbag observations of decision tree
}
\description{
Get inbag data of decision tree
}
\author{
Lea Louisa Kronziel, M.Sc.
}
