% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_observations_node.R
\name{get_observations_node}
\alias{get_observations_node}
\title{Get number of observations of the inbag data that reach respective node}
\usage{
get_observations_node(
  tree_info_df,
  train_data_df,
  rf_list,
  tree_number,
  node_id
)
}
\arguments{
\item{tree_info_df}{Data frame containing information about the structure of the decision tree, which is built like a "treeInfo()" data frame from the package "ranger"}

\item{train_data_df}{Data frame of the training data with which the random forest was trained}

\item{rf_list}{Random forest, which is built like the one you get from ranger()}

\item{tree_number}{Number of the decision tree of the rf_list to be displayed}

\item{node_id}{Node ID of the node whose parent ID is to be determined}
}
\value{
Character with the number of observations reaching this node
}
\description{
Get number of observations of the inbag data that reach respective node
}
\author{
Lea Louisa Kronziel, M.Sc.
}
