% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_parent_id.R
\name{get_parent_id}
\alias{get_parent_id}
\title{Get ID of parent node}
\usage{
get_parent_id(tree_info_df, node_id)
}
\arguments{
\item{tree_info_df}{Data frame containing information about the structure of the decision tree, which is built like a "treeInfo()" data frame from the package "ranger"}

\item{node_id}{Node ID of the node whose parent ID is to be determined}
}
\value{
Parent node ID (number)
}
\description{
Get ID of parent node
}
\author{
Lea Louisa Kronziel, M.Sc.
}
