% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_prediction_terminal_node.R
\name{get_prediction_terminal_node}
\alias{get_prediction_terminal_node}
\title{Get prediction of the node}
\usage{
get_prediction_terminal_node(
  tree_info_df,
  train_data_df,
  rf_list,
  dependent_var,
  tree_number,
  node_id
)
}
\arguments{
\item{tree_info_df}{Data frame containing information about the structure of the decision tree, which is built like a "treeInfo()" data frame from the package "ranger"}

\item{train_data_df}{Data frame of the training data with which the random forest was trained}

\item{rf_list}{Random forest, which is built like the one you get from ranger()}

\item{dependent_var}{Name of the column of the dependent variable in training data}

\item{tree_number}{Number of the decision tree of the rf_list to be displayed}

\item{node_id}{Node ID of the node whose parent ID is to be determined}
}
\value{
Character with the prediction of the node
}
\description{
Get prediction of the node
}
