% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_splitted_data.R
\name{get_splitted_data}
\alias{get_splitted_data}
\title{Get for each node which observations of the inbag data reach it}
\usage{
get_splitted_data(tree_info_df, inbag_data_df, rf_list, tree_number)
}
\arguments{
\item{tree_info_df}{Data frame containing information about the structure of the decision tree, which is built like a "treeInfo()" data frame from the package "ranger"}

\item{inbag_data_df}{Data frame of the inbag data with which the random forest was trained}

\item{rf_list}{Random forest, which is built like the one you get from ranger()}

\item{tree_number}{Number of the decision tree of the rf_list to be displayed}
}
\value{
List containing for each node the inbag data reaching this node
}
\description{
Get for each node which observations of the inbag data reach it
}
\author{
Laabs, Lea Louisa Kronziel, M.Sc.
}
