% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_uncertainty_node.R
\name{get_uncertainty_node}
\alias{get_uncertainty_node}
\title{Get uncertainty quantification for terminal nodes (only for regression tree)}
\usage{
get_uncertainty_node(
  tree_info_df,
  rf_list,
  test_data_df,
  tree_number,
  node_id,
  dependent_var,
  show_coverage = TRUE
)
}
\arguments{
\item{tree_info_df}{Data frame containing information about the structure of the decision tree, which is built like a "treeInfo()" data frame from the package "ranger"}

\item{rf_list}{Random forest, which is built like the one you get from ranger()}

\item{test_data_df}{Data frame of the test data}

\item{tree_number}{Number of the decision tree of the rf_list to be displayed}

\item{node_id}{Node ID of the node whose parent ID is to be determined}

\item{dependent_var}{Name of the dependent variable used to create the forest}

\item{show_coverage}{Option to display marginal coverage}
}
\value{
Character with the number of observations reaching this node
}
\description{
Get uncertainty quantification for terminal nodes (only for regression tree)
}
\author{
Lea Louisa Kronziel, M.Sc.
}
