% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/timbr-node.R
\name{getNodeFunction}
\alias{getNodeFunction}
\title{Node membership function}
\usage{
getNodeFunction(x, n, species)
}
\arguments{
\item{x}{a \link{timbr} object}

\item{n}{an integer corresponding to the node sequence number}

\item{species}{the species of timbr, e.g (randomForest, GBM)}
}
\value{
A function that returns a boolean value for the logical condition
represented by a node.
}
\description{
Node membership function
}

