% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/lumber.R
\name{harvest}
\alias{harvest}
\title{Process timbr into lumber}
\usage{
harvest(x)
}
\arguments{
\item{x}{A \code{\link{timbr}} object}
}
\value{
A lumber object
}
\description{
Process timbr into lumber
}
\details{
Process a timbr object, which a special class consisting of a
matrix and model metadata, into a lumber object. Whereas the timbr object
describes what a tree looks like, the lumber object describes what the tree
can do. It provides facilities for prediction and interpretation.
}

