% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/timbr-utility.R
\name{imputeMissing}
\alias{imputeMissing}
\title{Impute missing values}
\usage{
imputeMissing(x)
}
\arguments{
\item{x}{a data.frame object}
}
\value{
a \code{data.frame} object with an appended \code{imputed} class and
attribute.
}
\description{
Naive impuation using mode for factors and median for numerics. The function
returns a modified \code{data.frame} where missing values have been replaced
with modes for factors and medians for numerics. The returned object has the
imputed values added as the attribute "imputed".
}

