% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/forest.R
\name{is_forest}
\alias{is_forest}
\title{Test if an object is a forest}
\usage{
is_forest(x)
}
\arguments{
\item{x}{An object.}
}
\value{
`TRUE` if an object inherits from `forest` class.
}
\description{
Test if an object is a forest
}
