% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/lumber.R
\name{lumber}
\alias{lumber}
\title{Lumber constructor}
\usage{
lumber(x)
}
\arguments{
\item{x}{A list of \link{nodes}}
}
\value{
A processed timbr object, or lumber.
}
\description{
Lumber constructor
}
\details{
The \code{lumber} object is created after harvesting
\code{\link{timbr}}. See \link{harvest} for details on processing timbr.
}

