% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/lumberYard.R
\name{lumberYard}
\alias{lumberYard}
\title{Create a lumberYard from a decistion tree producint object}
\usage{
lumberYard(object, ...)
}
\arguments{
\item{object}{an object produced by a decision tree model aglorithm such as
randomForest or GBM. Any model such model can be used with lumberYard
provided there is a timbr and species specification for the model.}
}
\value{
a lumberYard object
}
\description{
lumberYard wraps the timbr conversion and harvesting process into one, simple
function call. It is the object by which decision tree models can be broken
down into smaller pieces called nodes or rules.
}

