% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deprecated.R
\name{map_forest}
\alias{map_forest}
\title{Apply a function hierarchically to a forest}
\usage{
map_forest(.x, .f, ..., .climb = FALSE)
}
\arguments{
\item{.x}{A forest}

\item{.f}{A function, formula, or vector (not necessarily atomic).}

\item{...}{Additional arguments passed on to the mapped function.}

\item{.climb}{Climbing or descending?}
}
\value{
A forest.
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#defunct}{\figure{lifecycle-defunct.svg}{options: alt='[Defunct]'}}}{\strong{[Defunct]}}

\code{map_forest()} has been removed. Please use \code{\link[=traverse]{traverse()}} instead.
}
