% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/timbr-utility.R
\name{massageNodes}
\alias{massageNodes}
\title{Massage nodes matrix}
\usage{
massageNodes(x, size = 50, drop.dups = TRUE, sparse = FALSE)
}
\arguments{
\item{x}{A matrix of binary node indicators.}

\item{size}{the minimum observations for any node.}

\item{drop.dups}{a boolean value for whether duplicate columns should be
zeroed out from the matrix.}

\item{sparse}{a boolean value for whether the returned matrix should be
sparse. Sparse matrices save memory.}
}
\value{
a matrix the same dimension as the input matrix \code{x} or a
sparse Matrix if \code{sparse} is \code{TRUE}.
}
\description{
Post process a matrix of nodes to eliminate those with insufficient counts
as well as duplicate nodes. If either the zeroes or ones of a column have
fewer than \code{size} observations, those columns will be zeroed out. If
\code{drop.dups} is set to \code{TRUE} then duplicate columns will also be
zeroed out.
}

