% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mcv-data.R
\docType{data}
\name{mcv.data}
\alias{mcv.data}
\title{Collaborative Cross QTL data for mean red blood cell volume (MCV)}
\format{A list of inputs formatted for use with TIMBR.}
\source{
MCV data from Kelada et al. (2012); recomputed diplotype state probabilities using similar methods
}
\usage{
data(mcv.data)
}
\description{
Data for mean red blood cell volume (MCV) QTL idenfied in Kelada et al. (2012).
}
\examples{
data(mcv.data)
str(mcv.data)
TIMBR.results <- TIMBR(mcv.data$y, mcv.data$prior.D, mcv.data$prior.M$crp)
}
\references{
Kelada et al. (2012) G3 2:157-165
(\href{http://www.ncbi.nlm.nih.gov/pubmed/22384394}{PubMed})
}
\keyword{datasets}
