% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/purrr.R
\name{modify.forest}
\alias{modify.forest}
\title{Modify a forest hierarchically}
\usage{
\method{modify}{forest}(.x, .f, ..., .climb = FALSE)
}
\arguments{
\item{.x}{A forest}

\item{.f}{A function, formula, or vector (not necessarily atomic).}

\item{...}{Additional arguments passed on to the mapped function.}

\item{.climb}{Climbing or descending?}
}
\value{
A forest.
}
\description{
Modifies nodes of a forest in the climbing or descending direction.
}
