% Generated by roxygen2 (4.0.2): do not edit by hand
\name{node}
\alias{node}
\title{Node class constructor}
\usage{
node(x, n, species, parent = NULL)
}
\arguments{
\item{x}{a \link{timbr} object}

\item{n}{an integer corresponding to the node sequence number}

\item{species}{the species of timbr, e.g (randomForest, GBM)}

\item{parent}{the sequence number if the parent node}
}
\value{
a node object consisting of node text, functions, and other metadata
}
\description{
Node class constructor
}

