% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TIMBR_source.R
\name{plot.hap.effects}
\alias{plot.hap.effects}
\title{Plot Haplotype Effects}
\usage{
\method{plot}{hap.effects}(TIMBR.output, file.name = "plot.png")
}
\arguments{
\item{TIMBR.output}{result object from the TIMBR function}

\item{file.name}{filename for the plot, saved in working directory}
}
\value{
a PNG plot of posterior haplotype effect densities
}
\description{
Plot posterior haplotype effect densities from TIMBR output
}
\examples{
#example data
data(mcv.data)
str(mcv.data)

#call TIMBR using CRP
results <- TIMBR(mcv.data$y, mcv.data$prior.D, mcv.data$prior.M$crp)

#plot haplotype effects
plot.hap.effects(results)

}
