% Generated by roxygen2 (4.0.2): do not edit by hand
\name{predict.lumberYard}
\alias{predict.lumberYard}
\title{Predict function for lumberYard objects}
\usage{
\method{predict}{lumberYard}(x, newdata, i = NULL, type = "Terminal")
}
\arguments{
\item{x}{a \code{\link{lumberYard}} object}

\item{newdata}{a dataset for which predictions are to be returned. Should
have the same columns and order as the dataset used to train the model.}

\item{i}{an optional numeric vector indicating a subset of nodeIDs to predict}

\item{type}{several forms of prediction are provided:
\enumerate{
 \item 'Terminal' - Used to predict the terminal node values.
 \item 'All' - Returns a matrix of 0s and 1s consisting of node membership.
 Note that all of the nodes are returned including internal ones.
 \item 'IDs' - Rather than the terminal node prediction, return the terminal
 node ID. Can be used to create different prediction values than those stored
 by the decistion tree model object used to create the lumberYard.
 \item 'Nodes' - When passed \code{i}, returns binary membership for only
 those requested nodes.}}
}
\description{
A generic predict method has been implemented for lumberYard objects. Several
types of prediction are available. See params for details.
}

