% Generated by roxygen2 (4.0.2): do not edit by hand
\name{print.node}
\alias{print.node}
\title{Print node}
\usage{
\method{print}{node}(node, newdata = NULL, y = NULL)
}
\arguments{
\item{node}{}

\item{newdata}{A dataset consisting of the same columns and order as the
dataset used to train the underlying decision tree.}

\item{y}{A response variable to be aggregated showing how the node performs}
}
\description{
Print the logical conditions or rules necessary to pass through a node. If
optional dataset and target variable are passed as well, the performance for
the node will be summarized in a table.
}
\details{
if \code{newdata} and \code{y} are passed into the function as well,
a table of summarized performance will be printed after the node text.
}

