% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/lumberYard.R
\name{printNodes}
\alias{printNodes}
\title{Print specific nodes}
\usage{
printNodes(x, nodes, newdata = NULL, y = NULL)
}
\arguments{
\item{x}{a \code{\link{lumberYard}} object}

\item{nodes}{a numeric vector of nodeIDs}

\item{newdata}{a dataset for which predictions are to be returned. Should
have the same columns and order as the dataset used to train the model.}

\item{y}{an optional numeric vector indicating a subset of nodeIDs to predict}
}
\value{
prints text representation to the console. Redirect using \code{sink}.
}
\description{
Print the text representation of node membership and optionally print target
variable statistics if provided a dataset and target variable.
}

