% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TIMBR_source.R
\name{TIMBR.biallelic.consistency}
\alias{TIMBR.biallelic.consistency}
\title{Calculate Biallelic Consistency Scores from TIMBR Output}
\usage{
TIMBR.biallelic.consistency(TIMBR.output, index = NULL, return.index = F,
  sort = T)
}
\arguments{
\item{TIMBR.output}{results object from the TIMBR function}

\item{index}{a pre-calculated index, as supplied by return.index=T}

\item{return.index}{optionally return the index used to compute the scores}

\item{sort}{optionally disable sorting of the consistency scores}
}
\value{
a named vector of consistency scores, or optionally a list that also contains the index
}
\description{
Calculates consistency scores for all possible biallelic contrasts from TIMBR output
}
\examples{
#example data
data(mcv.data)
str(mcv.data)

#call TIMBR using CRP
results <- TIMBR(mcv.data$y, mcv.data$prior.D, mcv.data$prior.M$crp)

#calculate biallelic consistency
TIMBR.biallelic.consistency(results)

}
