% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TIMBR_source.R
\name{TIMBR.consistent}
\alias{TIMBR.consistent}
\title{Construct consistent prior from existing list-type prior.M}
\usage{
TIMBR.consistent(prior.M, M.ID)
}
\arguments{
\item{prior.M}{prior.M object for use as TIMBR input, model.type must be list}

\item{M.ID}{string that denotes the specified biallelic contrast}
}
\value{
updated prior.M object that is consistent with M.ID
}
\description{
Updates prior.M to be consistent with biallelic contrast encoded by M.ID
}
\examples{
#example data
data(mcv.data)

#call TIMBR using CRP
results <- TIMBR(mcv.data$y, mcv.data$prior.D, mcv.data$prior.M$crp)

#approximate biallelic consistency BFs for all contrasts
head(TIMBR.approx(results))

#generate list object for CRP prior
prior.alpha <- list(type="gamma", shape=results$prior.M$prior.alpha.shape, rate=results$prior.M$prior.alpha.rate)
prior.M <- ewenss.exact(8, prior.alpha)

#specify biallelic contrast for consistency and update prior.M
M.ID <- "0,1,0,1,1,0,0,0"
prior.M <- TIMBR.consistent(prior.M, M.ID)

#analyze results, compare with approximate BF from earlier
results.consistent <- TIMBR(mcv.data$y, mcv.data$prior.D, prior.M)
results.consistent$ln.BF

}
