% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TIMBR_source.R
\name{TIMBR.plot.circos}
\alias{TIMBR.plot.circos}
\title{Circos Plot of Pairwise Partition Probabilities from TIMBR Object}
\usage{
TIMBR.plot.circos(TIMBR.object, file.path = NULL, plot.width = 480,
  plot.height = 480, post.summary = "mean", colors = c("blue",
  "white", "red"), color.res = 1000)
}
\arguments{
\item{TIMBR.object}{results object from the TIMBR function; or prior.M object of "list" type}

\item{file.path}{an optional file path for saving the plot as a PNG}

\item{plot.width}{PNG plot width}

\item{plot.height}{PNG plot height}

\item{post.summary}{posterior summary for relative effect size; select from c("mean", "median", "mode")}

\item{colors}{colors for heatmap of relative effect sizes}

\item{color.res}{number of colors to use for the heatmap color ramp}
}
\value{
circos plot of pairwise partition probabilities
}
\description{
Plots the probability that each pair of haplotypes is partitioned together on a circos plot; includes heatmap of relative effect sizes when using TIMBR results
}
\examples{
#example data
data(mcv.data)
str(mcv.data)

#call TIMBR using CRP
results <- TIMBR(mcv.data$y, mcv.data$prior.D, mcv.data$prior.M$crp)

#plot partition probabilities
TIMBR.plot.circos(results)

}
