% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TIMBR_source.R
\name{TIMBR.plot}
\alias{TIMBR.plot}
\title{Plot Haplotype Effects from TIMBR Output}
\usage{
TIMBR.plot(TIMBR.output, file.name = "plot.png")
}
\arguments{
\item{TIMBR.output}{result object from the TIMBR function}

\item{file.name}{filename for the plot, saved in working directory}
}
\value{
a PNG plot of posterior haplotype effect densities
}
\description{
Plots posterior haplotype effect densities from TIMBR output
}
\examples{
#example data
data(mcv.data)
str(mcv.data)

#call TIMBR using CRP
results <- TIMBR(mcv.data$y, mcv.data$prior.D, mcv.data$prior.M$crp)

#plot haplotype effects
plot.hap.effects(results)

}
