% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TIMBR_source.R
\name{TIMBR}
\alias{TIMBR}
\title{Tree-based Inference of Multiallelism via Bayesian Regression}
\usage{
TIMBR(
  y,
  prior.D,
  prior.M,
  prior.phi.v = 2,
  samples = 10000,
  Z = NULL,
  W = NULL,
  calc.lnBF = T,
  samples.ml = 1000,
  verbose = T,
  stop.on.error = F
)
}
\arguments{
\item{y}{vector of phenotype values for each strain}

\item{prior.D}{list of inputs for the prior distribution of strain diplotype states; see data(mcv.data) for an example}

\item{prior.M}{list of inputs for the prior distribution of the allelic series model; see data(mcv.data) for examples}

\item{prior.phi.v}{degrees of freedom for the half-t prior distribution on the variance component}

\item{samples}{number of samples to draw from the full posterior}

\item{Z}{design matrix for intercept and covariates; first column must be a vector of ones, which is the default}

\item{W}{vector of replicates for each strain; one replicate per strain by default}

\item{calc.lnBF}{option to calculate the lnBF, which is often computationally demanding}

\item{samples.ml}{number of samples to draw from the conditional posterior (if necessary) when calc.lnBF=T}

\item{verbose}{optionally report function progress}

\item{stop.on.error}{stop function if error is encountered when using 'integrate'. errors related to roundoff and small values may occur during edge cases}
}
\value{
a list of input parameters, posterior samples and marginal densities, and the marginal likelihood
}
\description{
Posterior samples and Bayes Factors using the TIMBR model
}
\examples{
#example data
data(mcv.data)
str(mcv.data)

#call TIMBR using CRP
results <- TIMBR(mcv.data$y, mcv.data$prior.D, mcv.data$prior.M$crp)

#report the Bayes Factor
results$ln.BF

#report posterior probabilities for the top allelic series models
head(results$p.M.given.y)

#report mean posterior haplotype effects
colMeans(results$post.hap.effects)

}
