% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/traverse.R
\name{traverse}
\alias{traverse}
\title{Apply a function hierarchically to a forest}
\usage{
traverse(.x, .f, ..., .climb = FALSE)
}
\arguments{
\item{.x}{A forest}

\item{.f}{A function, formula, or vector (not necessarily atomic).}

\item{...}{Additional arguments passed on to the mapped function.}

\item{.climb}{Climbing or descending?}
}
\value{
A forest.
}
\description{
Apply a function hierarchically to a forest in the climbing or descending direction.
}
