% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tree_to_text.R
\name{tree_to_text}
\alias{tree_to_text}
\title{Get latex code in the format of the package "forest" for the plot of the decision tree}
\usage{
tree_to_text(
  node_id,
  tree_info_df,
  train_data_df,
  test_data_df,
  cal_data_df,
  rf_list,
  tree_number,
  dependent_var,
  threshold,
  significance_level,
  interval_type,
  direction,
  show_sample_size,
  show_prediction_nodes,
  show_uncertainty,
  show_coverage,
  show_intervalwidth,
  show_cpd,
  cpd_plot_width,
  show_point_prediction,
  show_prediction_interval,
  vert_sep,
  hor_sep,
  colors,
  work_dir
)
}
\arguments{
\item{node_id}{Node ID of the node whose parent ID is to be determined}

\item{tree_info_df}{Data frame containing information about the structure of the decision tree, which is built like a "treeInfo()" data frame from the package "ranger"}

\item{train_data_df}{Data frame of the training data with which the random forest was trained}

\item{test_data_df}{Data frame of the test data (only needed, if show_coverage = TRUE)}

\item{cal_data_df}{Data frame of the calibration data (only needed, if show_cpd = TRUE)}

\item{rf_list}{Random forest, which is built like the one you get from ranger()}

\item{tree_number}{Number of the decision tree of the rf_list to be displayed}

\item{dependent_var}{Name of the column of the dependent variable in training data}

\item{threshold}{Vector of numerical threshhold(s) for which the probability of reaching this threshold is to be displayed in the plot; in addition, the probability of reaching at least the threshold is returned. Set to NULL if not used.}

\item{significance_level}{Level of uncertainty that should be reached by calibration, should be between 0 and 1.}

\item{interval_type}{Type of interval, choose either two-tailed or one-tailed}

\item{direction}{Direction of one-tailed interval, choose either left-tailed or right-tailed. Set to NULL if not used.}

\item{show_sample_size}{Option to display percentage of observations that reach nodes during training (TRUE or FALSE)}

\item{show_prediction_nodes}{Option to display prediction in all nodes (TRUE or FALSE)}

\item{show_uncertainty}{Option to display uncertainty quantification in terminal nodes (for now only available for regression)}

\item{show_coverage}{Option to display marginal coverage (only in combination with show_uncertainty = TRUE)}

\item{show_intervalwidth}{Option to display interval width uncertainty quantification in terminal nodes (only in combination with show_uncertainty = TRUE)}

\item{show_cpd}{Option to display cummulative prediction distribution (cpd) for each terminal node below each terminal node}

\item{cpd_plot_width}{Width of cpd plot}

\item{show_point_prediction}{Add dashed line in cpd plot for point prediction of terminal nodes}

\item{show_prediction_interval}{Add prediction interval in cpd plot}

\item{vert_sep}{Vertical spacing of nodes in mm (parameter from Latex package "forest")}

\item{hor_sep}{Horizontal spacing of nodes in mm (parameter from Latex package "forest")}

\item{colors}{Vector with color names with one entry for each node, so for each row in tree_info_df}

\item{work_dir}{Path where plot should be saved}
}
\value{
Character pasted Latex code for the plot with the Latex package "forest"
}
\description{
Get latex code in the format of the package "forest" for the plot of the decision tree
}
\author{
Lea Louisa Kronziel, M.Sc.
}
