% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TIMBR_source.R
\name{additive.design}
\alias{additive.design}
\title{Create an additive genetic design matrix}
\usage{
additive.design(J, type)
}
\arguments{
\item{J}{number of haplotypes}

\item{type}{format of the diplotype matrix, currently supports "happy" and "qtl2"}
}
\value{
matrix of (half) haplotype dosages corresponding to each diplotype state
}
\description{
Returns the additive design matrix for the number of specified haplotypes, in the specified format.
}
\examples{
additive.design(8, "happy")
additive.design(8, "qtl2")

}
