% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/forest.R
\name{as_forest}
\alias{as_forest}
\alias{as_forest.rowwise_df}
\alias{as_forest.grouped_df}
\title{Coerce to a forest}
\usage{
as_forest(x, ...)

\method{as_forest}{rowwise_df}(x, ...)

\method{as_forest}{grouped_df}(x, ...)
}
\arguments{
\item{x}{An object.}

\item{...}{Unused, for extensibility.}
}
\value{
A forest.
}
\description{
Coerce to a forest
}
