% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TIMBR_source.R
\name{calc.concentration.prior}
\alias{calc.concentration.prior}
\title{Calculate hyperparameters for Chinese restaurant process with gamma prior distribution on the concentration parameter}
\usage{
calc.concentration.prior(J, p.1.target, p.J.target)
}
\arguments{
\item{J}{number of customers (haplotypes) to be partitioned}

\item{p.1.target}{prior probability of 1 partition}

\item{p.J.target}{prior probability of J partitions}
}
\value{
vector of c(shape, rate) hyperparamters for the gamma distribution
}
\description{
Uses optimization to find hyperparameters that approximately yield the specified prior probabilities. May not work well when J is large or target prior probabilities are extreme.
}
\examples{
calc.concentration.prior(8, 0.05, 0.001)

}
