% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/forest.R
\name{children}
\alias{children}
\title{Children of the forest}
\usage{
children(data, name = NULL)
}
\arguments{
\item{data}{A forest.}

\item{name}{`NULL` (default) or a scalar character specifying the node name
of child nodes.}
}
\value{
A forest.
}
\description{
Convert a forest into a forest consisting of its child nodes.
}
