% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dplyr.R
\name{dplyr}
\alias{dplyr}
\alias{mutate.forest}
\alias{summarise.forest}
\alias{select.forest}
\alias{relocate.forest}
\alias{rows_update.forest}
\alias{rows_patch.forest}
\title{dplyr methods for forest objects}
\usage{
\method{mutate}{forest}(.data, ...)

\method{summarise}{forest}(.data, ..., .node = NULL)

\method{select}{forest}(.data, ...)

\method{relocate}{forest}(.data, ...)

\method{rows_update}{forest}(x, y, by = NULL, ...)

\method{rows_patch}{forest}(x, y, by = NULL, ...)
}
\arguments{
\item{.data}{A forest.}

\item{...}{Other arguments.}

\item{.node}{`NULL` (default) or a vector to create new nodes.}

\item{x}{A forest.}

\item{y}{A data frame.}

\item{by}{An unnamed character vector giving the key columns.}
}
\value{
A forest.
}
\description{
dplyr methods for forest objects.
}
