% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TIMBR_source.R
\name{ewenss.sampler}
\alias{ewenss.sampler}
\title{Ewens's sampling formula with gamma prior distribution on the concentration parameter}
\usage{
ewenss.sampler(samples, J, prior.alpha.a, prior.alpha.b, trees = NULL,
  verbose = T)
}
\arguments{
\item{samples}{number of samples}

\item{J}{number of leaves (haplotypes) to be partitioned, superceded by trees argument if specified}

\item{prior.alpha.a}{shape hyperparameter for the gamma distribution}

\item{trees}{optional user-specified tree(s) of class "phylo" ("multiPhylo"), detailed in the 'ape' package}

\item{verbose}{optionally report function progress}

\item{prior.alpha.a}{rate hyperparameter for the gamma distribution}
}
\value{
list of allelic series IDs and probabilities, formatted as prior.M argument for TIMBR function
}
\description{
Sample allelic series (paritions) from Ewen's sampling formula, optionally informed by user-specified tree(s). Branch lengths for specified trees must be in coalescent units for appropriate inference.
}
\examples{
#specifying hyperparameters using calc.concentration.prior
hyperparam <- calc.concentration.prior(8, 0.05, 0.001)

#running the sampler without user-specified trees; compare with target prior probabilities
prior.M <- ewenss.sampler(100000, 8, hyperparam[1], hyperparam[2])
prior.M$probs[prior.M$M.IDs=="0,0,0,0,0,0,0,0"]
prior.M$probs[prior.M$M.IDs=="0,1,2,3,4,5,6,7"]

#running the sampler with a user-specified tree; compare with tree structure
trees <- ape::rcoal(8, LETTERS[1:8])
ape::plot.phylo(trees)
prior.M <- ewenss.sampler(100000, 8, hyperparam[1], hyperparam[2], trees)
head(prior.M$M.IDs)
head(prior.M$probs)

}
