% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_tree.R
\name{generate_tree}
\alias{generate_tree}
\title{Generate most representative tree for a random forest}
\usage{
generate_tree(
  rf,
  metric = "weighted splitting variables",
  train_data,
  test_data = NULL,
  dependent_varname,
  importance.mode = FALSE,
  imp.num.var = NULL,
  ...
)
}
\arguments{
\item{rf}{Object of class \code{ranger} used with \code{write.forest = TRUE} to generate tree for.}

\item{metric}{Specification of the tree metric. Available are "splitting variables",
"weighted splitting variables", "terminal nodes" and "prediction".}

\item{train_data}{Data set for training of artificial representative tree}

\item{test_data}{Additional data set comparable to the data set \code{rf} was build on.}

\item{dependent_varname}{Name of the dependent variable used to create the forest}

\item{importance.mode}{If TRUE variable importance measures will be used to prioritize next split in tree generation.
Improves speed. Variable importance values have to be included in ranger object.}

\item{imp.num.var}{Number of variables to be pre selected based on importance values. If "automatic" the Boruta
variable selection algorithm from Kursa et al. (2010) is used (could be time consuming).
Insert a numeric value, if you want to define the number on your own.}

\item{...}{Further parameters passed on to Boruta (e.g. pValue)}
}
\value{
\item{\code{rep.trees}}{\code{ranger} object containing the artificial most representative tree}
}
\description{
\code{generate_tree} uses pair-wise dissimilarity of trees in a random
forest trained with \code{ranger} to generate an artificail most representative tree
which is not part aof the original ensemble
}
\examples{
require(ranger)
require(timbR)

## Train random forest with ranger
rg.iris <- ranger(Species ~ .,
                  data = iris,
                  write.forest=TRUE,
                  num.trees = 10,
                  importance = "permutation"
                  )

## Calculate pair-wise distances for all trees
rep_tree <- generate_tree(rf = rg.iris, metric = "splitting variables", train_data = iris, dependent_varname = "Species", importance.mode = TRUE, imp.num.var = 1)

}
\author{
Bjoern-Hergen Laabs, M.Sc.
}
