% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_distance_score.R
\name{get_distance_score}
\alias{get_distance_score}
\title{\code{get_distance_score} calculated tree based distance score for a single tree to a random forest built with \code{ranger}.}
\usage{
get_distance_score(rf, dist_val_rf, dist_val_tree, metric, test_data = NULL)
}
\arguments{
\item{rf}{Object of class \code{ranger} used with \code{write.forest = TRUE}.}

\item{dist_val_rf}{Distance values for the random forest created with get_distance_values()}

\item{dist_val_tree}{Distance values for the tree created with get_distance_values()}

\item{metric}{Specification of distance (dissimilarity) metric. Available are "splitting variables",
"weighted splitting variables", and "prediction".}

\item{test_data}{Additional data set comparable to the data set \code{rf} was build on. Only needed for metric "prediction".}
}
\value{
mean distance of tree to random forest
}
\description{
\code{get_distance_score} calculated tree based distance score for a single tree to a random forest built with \code{ranger}.
}
\author{
Lea Louisa Kronziel, M.sc.
}
