% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/timbr-node.R
\name{getNodeText}
\alias{getNodeText}
\title{Node membership text}
\usage{
getNodeText(x, n, species)
}
\arguments{
\item{x}{a \link{timbr} object}

\item{n}{an integer corresponding to the node sequence number}

\item{species}{the species of timbr, e.g (randomForest, GBM)}
}
\value{
Returns the text representing the logical condition of the parent
node
}
\description{
Node membership text
}

