% Generated by roxygen2 (4.0.2): do not edit by hand
\name{node.print.func}
\alias{node.print.func}
\title{Target variable statistics specification}
\usage{
node.print.func(x)
}
\arguments{
\item{x}{target variable}
}
\value{
This function is used to aggregate a target variable over the levels
of a node. Returns the toal count, sum, and mean of the target variable.
}
\description{
Target variable statistics specification
}

