% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/node.R
\name{node}
\alias{node}
\alias{node_name}
\alias{node_value}
\alias{node_parent}
\title{Attributes of root nodes}
\usage{
node_name()

node_value()

node_parent()
}
\value{
A vector of names, values, or parents of root nodes.
}
\description{
Attributes of root nodes
}
