% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/timbr-package.R
\docType{package}
\name{timbr-package}
\alias{timbr}
\alias{timbr-package}
\title{timbr: Forest/Tree Data Frames}
\description{
Provides data frames for forest or tree data structures. You can create forest data structures from data frames and process them based on their hierarchies.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/UchidaMizuki/timbr}
  \item \url{https://uchidamizuki.github.io/timbr/}
  \item Report bugs at \url{https://github.com/UchidaMizuki/timbr/issues}
}

}
\author{
\strong{Maintainer}: Mizuki Uchida \email{uchidamizuki@vivaldi.net}

}
\keyword{internal}
