% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TIMBR_source.R, R/TIMBR_source_bk.R,
%   R/TIMBR_source_bk2.R, R/TIMBR_source_bk3.R
\name{TIMBR.approx}
\alias{TIMBR.approx}
\title{Approximate Bayes Factors from TIMBR Output}
\usage{
TIMBR.approx(TIMBR.output, type = "consistent")

TIMBR.approx(TIMBR.output, type = "consistent")

TIMBR.approx(TIMBR.output, type = "consistent")

TIMBR.approx(TIMBR.output, type = "consistent")
}
\arguments{
\item{TIMBR.output}{results object from the TIMBR function}

\item{type}{"all" - report BFs for all allelic series; "merge" - report BFs for biallelic series (i.e. merge analysis); "consistent" - report BFs for consistent merge analysis}

\item{TIMBR.output}{results object from the TIMBR function}

\item{type}{"all" - report BFs for all allelic series; "merge" - report BFs for biallelic series (i.e. merge analysis); "consistent" - report BFs for consistent merge analysis}

\item{TIMBR.output}{results object from the TIMBR function}

\item{type}{"all" - report BFs for all allelic series; "merge" - report BFs for biallelic series (i.e. merge analysis); "consistent" - report BFs for consistent merge analysis}

\item{TIMBR.output}{results object from the TIMBR function}

\item{type}{"all" - report BFs for all allelic series; "merge" - report BFs for biallelic series (i.e. merge analysis); "consistent" - report BFs for consistent merge analysis}
}
\value{
a named vector of approximate Bayes Factors; appoximations that underflow are not reported

a named vector of approximate Bayes Factors; appoximations that underflow are not reported

a named vector of approximate Bayes Factors; appoximations that underflow are not reported

a named vector of approximate Bayes Factors; appoximations that underflow are not reported
}
\description{
Approximate Bayes factors for various hypotheses from TIMBR output

Approximate Bayes factors for various hypotheses from TIMBR output

Approximate Bayes factors for various hypotheses from TIMBR output

Approximate Bayes factors for various hypotheses from TIMBR output
}
\examples{
#example data
data(mcv.data)
str(mcv.data)

#call TIMBR using CRP
results <- TIMBR(mcv.data$y, mcv.data$prior.D, mcv.data$prior.M$crp)

#calculate biallelic consistency
TIMBR.approx(results)

#example data
data(mcv.data)
str(mcv.data)

#call TIMBR using CRP
results <- TIMBR(mcv.data$y, mcv.data$prior.D, mcv.data$prior.M$crp)

#calculate biallelic consistency
TIMBR.approx(results)

#example data
data(mcv.data)
str(mcv.data)

#call TIMBR using CRP
results <- TIMBR(mcv.data$y, mcv.data$prior.D, mcv.data$prior.M$crp)

#calculate biallelic consistency
TIMBR.approx(results)

#example data
data(mcv.data)
str(mcv.data)

#call TIMBR using CRP
results <- TIMBR(mcv.data$y, mcv.data$prior.D, mcv.data$prior.M$crp)

#calculate biallelic consistency
TIMBR.approx(results)

}
