% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TIMBR_source.R
\name{TIMBR.plot.haplotypes}
\alias{TIMBR.plot.haplotypes}
\title{Plot Haplotype Effects from TIMBR Output}
\usage{
TIMBR.plot.haplotypes(TIMBR.output, colors = NULL, file.path = NULL,
  plot.width = 960, plot.height = 480, TIMBR.output.bkgrd = NULL,
  colors.bkgrd = NULL, transparency = c(0.7, 0.4), hap.labels = NULL,
  x.lim = NULL, x.lab = NULL)
}
\arguments{
\item{TIMBR.output}{results object from the TIMBR function}

\item{colors}{an optional vector of colors for each haplotype density}

\item{file.path}{an optional file path for saving the plot as a PNG}

\item{plot.width}{PNG plot width}

\item{plot.height}{PNG plot height}

\item{TIMBR.output.bkgrd}{optional second results object for background plot}

\item{colors.bkgrd}{optional colors for background plot}

\item{transparency}{optional vector of transparencies for foreground and background plots, ignored if second results object is not specified}

\item{hap.labels}{optional labels for haplotypes}

\item{x.lim}{optional limits for x axis}

\item{x.lab}{optional label for x axis}
}
\value{
plot of the posterior haplotype effect densities
}
\description{
Plots posterior haplotype effect densities from TIMBR output
}
\examples{
#example data
data(mcv.data)
str(mcv.data)

#call TIMBR using CRP
results <- TIMBR(mcv.data$y, mcv.data$prior.D, mcv.data$prior.M$crp)

#plot haplotype effects
TIMBR.plot.haplotypes(results)

}
