% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TIMBR_source.R, R/TIMBR_source_bk.R,
%   R/TIMBR_source_bk2.R, R/TIMBR_source_bk3.R
\name{TIMBR.plot}
\alias{TIMBR.plot}
\title{Plot Haplotype Effects from TIMBR Output}
\usage{
TIMBR.plot(TIMBR.output, colors = NULL, file.path = NULL,
  plot.width = 960, plot.height = 480)

TIMBR.plot(TIMBR.output, colors = NULL, file.path = NULL,
  plot.width = 960, plot.height = 480)

TIMBR.plot(TIMBR.output, colors = NULL, file.path = NULL,
  plot.width = 960, plot.height = 480)

TIMBR.plot(TIMBR.output, colors = NULL, file.path = NULL,
  plot.width = 960, plot.height = 480)
}
\arguments{
\item{TIMBR.output}{results object from the TIMBR function}

\item{colors}{an optional vector of colors for each haplotype density}

\item{file.path}{an optional file path for saving the plot as a PNG}

\item{plot.width}{PNG plot width}

\item{plot.height}{PNG plot height}

\item{TIMBR.output}{results object from the TIMBR function}

\item{colors}{an optional vector of colors for each haplotype density}

\item{file.path}{an optional file path for saving the plot as a PNG}

\item{plot.width}{PNG plot width}

\item{plot.height}{PNG plot height}

\item{TIMBR.output}{results object from the TIMBR function}

\item{colors}{an optional vector of colors for each haplotype density}

\item{file.path}{an optional file path for saving the plot as a PNG}

\item{plot.width}{PNG plot width}

\item{plot.height}{PNG plot height}

\item{TIMBR.output}{results object from the TIMBR function}

\item{colors}{an optional vector of colors for each haplotype density}

\item{file.path}{an optional file path for saving the plot as a PNG}

\item{plot.width}{PNG plot width}

\item{plot.height}{PNG plot height}
}
\value{
plot of the posterior haplotype effect densities

plot of the posterior haplotype effect densities

plot of the posterior haplotype effect densities

plot of the posterior haplotype effect densities
}
\description{
Plots posterior haplotype effect densities from TIMBR output

Plots posterior haplotype effect densities from TIMBR output

Plots posterior haplotype effect densities from TIMBR output

Plots posterior haplotype effect densities from TIMBR output
}
\examples{
#example data
data(mcv.data)
str(mcv.data)

#call TIMBR using CRP
results <- TIMBR(mcv.data$y, mcv.data$prior.D, mcv.data$prior.M$crp)

#plot haplotype effects
TIMBR.plot(results)

#example data
data(mcv.data)
str(mcv.data)

#call TIMBR using CRP
results <- TIMBR(mcv.data$y, mcv.data$prior.D, mcv.data$prior.M$crp)

#plot haplotype effects
TIMBR.plot(results)

#example data
data(mcv.data)
str(mcv.data)

#call TIMBR using CRP
results <- TIMBR(mcv.data$y, mcv.data$prior.D, mcv.data$prior.M$crp)

#plot haplotype effects
TIMBR.plot(results)

#example data
data(mcv.data)
str(mcv.data)

#call TIMBR using CRP
results <- TIMBR(mcv.data$y, mcv.data$prior.D, mcv.data$prior.M$crp)

#plot haplotype effects
TIMBR.plot(results)

}
