% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TIMBR_package.R
\name{TIMBR}
\alias{TIMBR}
\title{Tree-based Inference of Multiallelism via Bayesian Regression}
\usage{
TIMBR(y, prior.D, prior.M, prior.v.b = 1, samples = 1000,
  samples.ml = 1000, Z = NULL, W = NULL, verbose = T)
}
\arguments{
\item{y}{phenotype vector}

\item{prior.D}{diplotype prior list}

\item{prior.M}{model prior list}
}
\value{
None
}
\description{
Posterior samples and Bayes Factors using the TIMBR model
}
\examples{
TIMBR(y, prior.D, prior.M)

}
