% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TIMBR_source.R
\name{TIMBR}
\alias{TIMBR}
\title{Tree-based Inference of Multiallelism via Bayesian Regression}
\usage{
TIMBR(y, prior.D, prior.M, prior.v.b = 1, samples = 1000,
  samples.ml = 1000, Z = NULL, W = NULL, verbose = T)
}
\arguments{
\item{y}{vector of phenotype values for each strain}

\item{prior.D}{list of inputs for the prior distribution of strain diplotype states; see data(mcv.data) for an example}

\item{prior.M}{list of inputs for the prior distribution of the allelic series model; see data(mcv.data) for examples}

\item{prior.v.b}{shape parameter for the beta prime prior distribution on the variance component}

\item{samples.ml}{number of samples to draw from the condtiional posterior (if necessary)}

\item{Z}{design matrix for covariates; first column must be a vector of ones, which is the default}

\item{W}{vector of replicates for each strain; one replicate per strain by default}

\item{verbose}{optionally report function progress}

\item{sample}{number of samples to draw from the full posterior}
}
\value{
returns a list of input parameters, posterior samples, and the marginal likelihood
}
\description{
Posterior samples and Bayes Factors using the TIMBR model
}
\examples{
#example data
data(mcv.data)
str(mcv.data)

#call TIMBR using CRP
results <- TIMBR(mcv.data$y, mcv.data$prior.D, mcv.data$prior.M$crp)

#calculate the Bayes Factor
results$ln.ml - results$ln.ml.null

#report posterior probabilities for the top allelic series models
head(rev(sort(table(results$post.M))))/results$samples

}
