% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tree_to_text.R
\name{tree_to_text}
\alias{tree_to_text}
\title{Get latex code in the format of the package "forest" for the plot of the decision tree}
\usage{
tree_to_text(
  node_id,
  tree_info_df,
  train_data_df,
  rf_list,
  tree_number,
  dependent_var,
  show_sample_size,
  show_prediction_nodes,
  vert_sep,
  hor_sep
)
}
\arguments{
\item{node_id}{Node ID of the node whose parent ID is to be determined}

\item{tree_info_df}{Data frame containing information about the structure of the decision tree, which is built like a "treeInfo()" data frame from the package "ranger"}

\item{train_data_df}{Data frame of the training data with which the random forest was trained}

\item{rf_list}{Random forest, which is built like the one you get from ranger()}

\item{tree_number}{Number of the decision tree of the rf_list to be displayed}

\item{dependent_var}{Name of the column of the dependent variable in training data}

\item{show_sample_size}{Option to display percentage of observations that reach nodes during training (TRUE or FALSE)}

\item{show_prediction_nodes}{Option to display prediction in all nodes (TRUE or FALSE)}

\item{vert_sep}{Vertical spacing of nodes in mm (parameter from Latex package "forest")}

\item{hor_sep}{Horizontal spacing of nodes in mm (parameter from Latex package "forest")}
}
\value{
Character pasted Latex code for the plot with the Latex package "forest"
}
\description{
Get latex code in the format of the package "forest" for the plot of the decision tree
}
